test: refactor mcp coverage and add controller resume smoke test
