+++ controller/test/e2e/features/agentgateway/mcp/testdata/multireplica.yaml
@@ -1,0 +1,62 @@
+---
+apiVersion: agentgateway.dev/v1alpha1
+kind: AgentgatewayParameters
+metadata:
+  name: mcp-multireplica
+  namespace: default
+spec:
+  deployment:
+    spec:
+      replicas: 2
+  rawConfig:
+    config:
+      session:
+        key: 000102030405060708090a0b0c0d0e0f000102030405060708090a0b0c0d0e0f
+---
+apiVersion: gateway.networking.k8s.io/v1
+kind: Gateway
+metadata:
+  name: mcp-multireplica-gateway
+  namespace: default
+spec:
+  gatewayClassName: agentgateway
+  infrastructure:
+    parametersRef:
+      group: agentgateway.dev
+      kind: AgentgatewayParameters
+      name: mcp-multireplica
+  listeners:
+    - protocol: HTTP
+      port: 80
+      name: http
+      allowedRoutes:
+        namespaces:
+          from: All
+---
+apiVersion: agentgateway.dev/v1alpha1
+kind: AgentgatewayBackend
+metadata:
+  name: mcp-multireplica-backend
+  namespace: default
+spec:
+  mcp:
+    targets:
+      - name: mcp-target
+        static:
+          host: mcp-website-fetcher.default.svc.cluster.local
+          port: 80
+          protocol: StreamableHTTP
+---
+apiVersion: gateway.networking.k8s.io/v1
+kind: HTTPRoute
+metadata:
+  name: mcp-multireplica-route
+  namespace: default
+spec:
+  parentRefs:
+    - name: mcp-multireplica-gateway
+  rules:
+    - backendRefs:
+        - name: mcp-multireplica-backend
+          group: agentgateway.dev
+          kind: AgentgatewayBackend
